recreated QuestSystem
The logic from scene1 with quest system was not good. Created a scene two, deleted old scripts and made new logic.

Player can now accept quest
Player can see that the quest is in progress if he interacts with the npc again.
Player can complete the quest at npc if the quest is fullfilling the quest conditions.

UI
ui is now managed from the questManager that is related to the quest and interactions with objects and npc's.

Added a interact text
Added text when a quest item is being picked up and with timer that is fading out and then destroying object.